+++ src/main/java/net/mohron/skyclaims/command/admin/CommandSetup.java
@@ -18,19 +18,17 @@
 
 package net.mohron.skyclaims.command.admin;
 
-import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.permissions.Permissions;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
-import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
-import org.spongepowered.api.service.permission.PermissionService;
 import org.spongepowered.api.text.Text;
 
-public class CommandSetup implements CommandExecutor {
-	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
+public class CommandSetup extends CommandBase {
+
 	public static final String HELP_TEXT = "used to assist in setting up the plugin";
 
 	public static CommandSpec commandSpec = CommandSpec.builder()