+++ src/main/java/net/mohron/skyclaims/command/user/CommandList.java
@@ -20,6 +20,7 @@ package net.mohron.skyclaims.command.user;
 
 import com.google.common.collect.Lists;
 import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.command.argument.Argument;
 import net.mohron.skyclaims.permissions.Permissions;
 import net.mohron.skyclaims.util.CommandUtil;
@@ -29,7 +30,6 @@ import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.args.GenericArguments;
-import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.entity.living.player.User;
@@ -43,9 +43,9 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.function.Consumer;
 
-public class CommandList implements CommandExecutor {
-	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
-	public static final String HELP_TEXT = "display a list of the current islands";
+public class CommandList extends CommandBase {
+
+	public static final String HELP_TEXT = "display a list of the current islands.";
 	private static final Text USER = Text.of("user");
 	private static final Text SORT = Text.of("sort");
 
@@ -70,8 +70,9 @@ public class CommandList implements CommandExecutor {
 	}
 
 	public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-		if (SkyClaims.islands.isEmpty())
+		if (SkyClaims.islands.isEmpty()) {
 			src.sendMessage(Text.of(TextColors.RED, "There are currently no islands!"));
+		}
 		List<Text> listText = Lists.newArrayList();
 		Player player = (src instanceof Player) ? (Player) src : null;
 		User user = args.<User>getOne(USER).orElse(null);
@@ -81,26 +82,31 @@ public class CommandList implements CommandExecutor {
 
 		SkyClaims.islands.values().stream()
 			.filter(i -> user == null || i.hasPermissions(user))
-			.sorted(Comparator.comparing(Island::getSortableName))
 			.sorted(sortType)
 			.forEach(island -> listText.add(Text.of(
 				getLocked(island),
 				island.getName().toBuilder()
 					.onHover(TextActions.showText(Text.of("Click here to view island info")))
-					.onClick(TextActions.executeCallback(CommandUtil.createCommandConsumer(src, "islandinfo", island.getUniqueId().toString(), createReturnConsumer(src)))),
-				(!island.isLocked() || ((player == null || island.hasPermissions(player)) || spawnOthers)) ? getClickableCoords(src, island) : getCoords(island)
+					.onClick(TextActions.executeCallback(
+						CommandUtil.createCommandConsumer(src, "islandinfo", island.getUniqueId().toString(), createReturnConsumer(src)))
+					),
+				(!island.isLocked() || ((player == null || island.hasPermissions(player)) || spawnOthers))
+					? getClickableCoords(src, island) : getCoords(island)
 			)));
 
-		if (listText.isEmpty())
+		if (listText.isEmpty()) {
 			listText.add(Text.of(TextColors.RED, "There are no islands to display!"));
+		}
 
-		if (!(src instanceof Player))
+		if (!(src instanceof Player)) {
 			listText.forEach(src::sendMessage);
-		else PaginationList.builder()
-			.title(Text.of(TextColors.AQUA, "Island List"))
-			.padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
-			.contents(listText)
-			.sendTo(src);
+		} else {
+			PaginationList.builder()
+				.title(Text.of(TextColors.AQUA, "Island List"))
+				.padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
+				.contents(listText)
+				.sendTo(src);
+		}
 
 		return CommandResult.success();
 	}
@@ -108,8 +114,10 @@ public class CommandList implements CommandExecutor {
 	private static Consumer<CommandSource> createReturnConsumer(CommandSource src) {
 		return consumer -> {
 			Text returnCommand = Text.builder().append(Text.of(
-					Text.NEW_LINE, TextColors.WHITE, "[", TextColors.AQUA, "Return to Island List", TextColors.WHITE, "]", Text.NEW_LINE))
-					.onClick(TextActions.executeCallback(CommandUtil.createCommandConsumer(src, "islandlist", "", null))).build();
+				Text.NEW_LINE, TextColors.WHITE, "[", TextColors.AQUA, "Return to Island List", TextColors.WHITE, "]", Text.NEW_LINE))
+				.onClick(TextActions.executeCallback(
+					CommandUtil.createCommandConsumer(src, "islandlist", "", null))
+				).build();
 			src.sendMessage(returnCommand);
 		};
 	}
@@ -128,7 +136,10 @@ public class CommandList implements CommandExecutor {
 
 	private Consumer<CommandSource> toggleLock(Island island) {
 		return src -> {
-			if (src instanceof Player && ((Player) src).getUniqueId().equals(island.getOwnerUniqueId()) && src.hasPermission(Permissions.COMMAND_LOCK) || src.hasPermission(Permissions.COMMAND_LOCK_OTHERS)) {
+			if (src instanceof Player
+				&& ((Player) src).getUniqueId().equals(island.getOwnerUniqueId())
+				&& src.hasPermission(Permissions.COMMAND_LOCK)
+				|| src.hasPermission(Permissions.COMMAND_LOCK_OTHERS)) {
 				island.setLocked(!island.isLocked());
 				src.sendMessage(Text.of(island.getName(), TextColors.GREEN, " is now ",
 					Text.builder((island.isLocked()) ? "LOCKED" : "UNLOCKED")