+++ src/main/java/net/mohron/skyclaims/permissions/Permissions.java
@@ -28,6 +28,7 @@ public class Permissions {
 	public static final String COMMAND_INVITE = "skyclaims.command.invite";
 	public static final String COMMAND_LIST = "skyclaims.command.list";
 	public static final String COMMAND_LOCK = "skyclaims.command.lock";
+	public static final String COMMAND_REGEN = "skyclaims.command.regen";
 	public static final String COMMAND_RESET = "skyclaims.command.reset";
 	public static final String COMMAND_SPAWN = "skyclaims.command.spawn";
 	public static final String COMMAND_SET_HOME = "skyclaims.command.sethome";